fixing horizontal alignment and spacing in menu widget
Now the line spacing is parametrized, and defaults to 20. Setting
it to zero will place one menu item right below the other. Setting
a negative value will make them overlap.

The horizontal alignment is also fixed, and now allows for 'center',
'left' and 'right' values. The 'topleft' property becomes really the
topleft value for the menu, making it easier to place in your window.